[AN] refactor: 검색 화면 구조 리팩토링 및 컴포넌트 분리 #795
*   [AN] refactor: SearchViewModel을 SearchDetailViewModel로 대체 및 패키지 이동
*   [AN] refactor: SearchResult, SearchRecent 스크린 파일명 변경 (Screen 접미사 추가)
*   [AN] refactor: SearchRecentItem 컴포넌트 분리 및 하위 컴포넌트 패키지 정리
