+++ android/app/src/main/java/com/onair/hearit/presentation/search/detail/SearchRecentScreen.kt
@@ -1,7 +1,6 @@
 package com.onair.hearit.presentation.search.detail
 
 import androidx.compose.foundation.background
-import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
@@ -10,24 +9,20 @@ import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.heightIn
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.items
 import androidx.compose.material3.CircularProgressIndicator
-import androidx.compose.material3.Icon
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.onair.hearit.R
+import com.onair.hearit.presentation.search.detail.component.SearchRecentItem
 import com.onair.hearit.presentation.theme.Gray2
-import com.onair.hearit.presentation.theme.Gray4
 import com.onair.hearit.presentation.theme.HearitBlack
 import com.onair.hearit.presentation.theme.HearitTypoGraphy
 import com.onair.hearit.presentation.util.noRippleClickable
@@ -65,7 +60,7 @@ fun SearchRecent(
                         items = keywords,
                         key = { it },
                     ) { keyword ->
-                        RecentSearchItem(
+                        SearchRecentItem(
                             keyword = keyword,
                             onClick = { onKeywordClick(keyword) },
                         )
@@ -112,39 +107,6 @@ fun RecentSearchHeader(
 }
 
 @Composable
-private fun RecentSearchItem(
-    keyword: String,
-    onClick: () -> Unit,
-    modifier: Modifier = Modifier,
-) {
-    Row(
-        modifier =
-            modifier
-                .fillMaxWidth()
-                .clickable(onClick = onClick)
-                .padding(horizontal = 20.dp, vertical = 8.dp),
-        horizontalArrangement = Arrangement.SpaceBetween,
-        verticalAlignment = Alignment.CenterVertically,
-    ) {
-        Text(
-            text = keyword,
-            style = HearitTypoGraphy.bodyLarge,
-            color = Gray4,
-            maxLines = 1,
-            overflow = TextOverflow.Ellipsis,
-            modifier = Modifier.weight(1f),
-        )
-
-        Icon(
-            painter = painterResource(R.drawable.ic_recent_keyword),
-            contentDescription = null,
-            tint = Gray2,
-            modifier = Modifier.size(20.dp),
-        )
-    }
-}
-
-@Composable
 private fun EmptyRecentSearch(modifier: Modifier = Modifier) {
     Box(
         modifier =
@@ -161,7 +123,6 @@ private fun EmptyRecentSearch(modifier: Modifier = Modifier) {
     }
 }
 
-// Preview
 @Preview(showBackground = true, backgroundColor = 0xFF000000)
 @Composable
 private fun RecentSearchScreenPreview() {
@@ -190,24 +151,5 @@ private fun EmptyRecentSearchScreenPreview() {
             onKeywordClick = {},
             onClearAll = {},
         )
-    }
-}
-
-@Preview(showBackground = true, backgroundColor = 0xFF000000)
-@Composable
-private fun RecentSearchItemPreview() {
-    MaterialTheme {
-        Box(
-            modifier =
-                Modifier
-                    .fillMaxWidth()
-                    .background(HearitBlack)
-                    .padding(16.dp),
-        ) {
-            RecentSearchItem(
-                keyword = "Android 클린 아키텍처",
-                onClick = {},
-            )
-        }
     }
 }
