+++ android/app/src/main/java/com/onair/hearit/presentation/search/detail/SearchResultScreen.kt
@@ -18,6 +18,7 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.onair.hearit.domain.model.Keyword
 import com.onair.hearit.domain.model.SearchedHearit
+import com.onair.hearit.presentation.search.detail.component.SearchResultItem
 import com.onair.hearit.presentation.theme.HearitBlack
 import kotlinx.collections.immutable.persistentListOf
 import kotlinx.coroutines.flow.distinctUntilChanged
