+++ android/app/src/main/java/com/onair/hearit/presentation/search/detail/component/SearchRecentItem.kt
@@ -1,0 +1,77 @@
+package com.onair.hearit.presentation.search.detail.component
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.material3.Icon
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.text.style.TextOverflow
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.dp
+import com.onair.hearit.R
+import com.onair.hearit.presentation.theme.Gray2
+import com.onair.hearit.presentation.theme.Gray4
+import com.onair.hearit.presentation.theme.HearitBlack
+import com.onair.hearit.presentation.theme.HearitTypoGraphy
+
+@Composable
+fun SearchRecentItem(
+    keyword: String,
+    onClick: () -> Unit,
+    modifier: Modifier = Modifier,
+) {
+    Row(
+        modifier =
+            modifier
+                .fillMaxWidth()
+                .clickable(onClick = onClick)
+                .padding(horizontal = 20.dp, vertical = 8.dp),
+        horizontalArrangement = Arrangement.SpaceBetween,
+        verticalAlignment = Alignment.CenterVertically,
+    ) {
+        Text(
+            text = keyword,
+            style = HearitTypoGraphy.bodyLarge,
+            color = Gray4,
+            maxLines = 1,
+            overflow = TextOverflow.Ellipsis,
+            modifier = Modifier.weight(1f),
+        )
+
+        Icon(
+            painter = painterResource(R.drawable.ic_recent_keyword),
+            contentDescription = null,
+            tint = Gray2,
+            modifier = Modifier.size(20.dp),
+        )
+    }
+}
+
+@Preview(showBackground = true, backgroundColor = 0xFF000000)
+@Composable
+private fun SearchRecentItemPreview() {
+    MaterialTheme {
+        Box(
+            modifier =
+                Modifier
+                    .fillMaxWidth()
+                    .background(HearitBlack)
+                    .padding(16.dp),
+        ) {
+            SearchRecentItem(
+                keyword = "Android 클린 아키텍처",
+                onClick = {},
+            )
+        }
+    }
+}
